Show percentiles for subsets of global data
Before this update, the plot that showed percentiles of scores and
number of shows per year only did so using the entire globalDS dataset.
This update changes this so that the percentiles are shown for the
category chosen by the user.  That it, percentiles can be shown for
shows only with a certain type, or only a certain source, or only a
certain genre.